Added pgtype flag to CLI arguments.
This will change the go types in the generated files to types from the
`pgtype` module. Those types are natively supported by the `pgx` module.

* `pg-type` flag is an enum that can be `std`, 'pointer', `pgtype`, or
`pgtype-full`.
This lets users decide how they use `pgtype` in their projects;
* moved all the code related to postgres types in
`loaders/postgrestypes`. This was mandatory to avoid a cyclic dependency
as the enum is used in `internal` and `loaders` is used in `internal`;
* added `pgtype` import in `xo_package.go.tpl`. This is the only change
required in the templates to support `pgtype` and `pgx` — Gendal does
not care about SQL drivers: to support `pgx` we only need to support
`pgtype`.

https://github.com/jackc/pgtype